fixing codeclimate
