+++ modules/claims_api/app/controllers/claims_api/docs/v0/api_controller.rb
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 
-require_dependency 'claims_api/form_526_model_swagger'
-require_dependency 'claims_api/form_0966_model_swagger'
-require_dependency 'claims_api/form_2122_model_swagger'
 require_dependency 'claims_api/form_526_v0_controller_swagger'
 require_dependency 'claims_api/form_0966_v0_controller_swagger'
 require_dependency 'claims_api/form_2122_v0_controller_swagger'
@@ -10,11 +7,7 @@ require_dependency 'claims_api/form_2122_v0_controller_swagger'
 module ClaimsApi
   module Docs
     module V0
-      class ApiController < ::ApplicationController
-        skip_before_action(:authenticate)
-        include Swagger::Blocks
-
-        # A list of all classes that have swagger_* declarations.
+      class ApiController < ClaimsApi::Docs::ApiController
         SWAGGERED_CLASSES = [
           ClaimsApi::ClaimsModelSwagger,
           ClaimsApi::Form526ModelSwagger,
